chore (refs T22614): move store & lib for User to client/js
