+++ client/js/bundles/procedure/administrationEdit.js
@@ -19,7 +19,7 @@ import DpEmailList from '@DpJs/components/procedure/basicSettings/DpEmailList'
 import dpValidate from '@DpJs/lib/core/validation/dpValidate'
 import DPWizard from '@DpJs/lib/procedure/DPWizard'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import UrlPreview from '@DemosPlanUserBundle/lib/UrlPreview'
+import UrlPreview from '@DpJs/lib/user/UrlPreview'
 
 const components = { DpBasicSettings, DpEmailList, DpDateRangePicker }
