+++ client/js/bundles/user/editOrga.js
@@ -15,7 +15,7 @@ import DpAccordion from '@DpJs/components/core/DpAccordion'
 import DpEditor from '@DpJs/components/core/DpEditor/DpEditor'
 import dpValidate from '@DpJs/lib/core/validation/dpValidate'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import UrlPreview from '@DemosPlanUserBundle/lib/UrlPreview'
+import UrlPreview from '@DpJs/lib/user/UrlPreview'
 
 const components = {
   DpAccordion,
