+++ client/js/bundles/user/listUser.js
@@ -15,7 +15,7 @@ import DpCreateItem from '@DpJs/components/user/DpCreateItem'
 import DpUserList from '@DpJs/components/user/DpUserList'
 import DpUserListExtended from '@DpJs/components/user/DpUserListExtended'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
-import UserFormFields from '@DemosPlanUserBundle/store/UserFormFields'
+import UserFormFields from '@DpJs/store/user/UserFormFields'
 
 const stores = {
   UserFormFields
